shifted extract_oco2_files to oco2.py
