+++ acrg/satellite/oco2.py
@@ -49,7 +49,7 @@ from .barometric import pressure_at_height
 from acrg.countrymask import domain_volume
 from acrg.config.paths import Paths
 
-from acrg.satellite.common import extract_files_oco2, name_pressure_file, name_pressure_match,apply_filter,coord_order,define_pressure_levels,distance_lat,distance_lon,output_name,output,add_coords,ds_check_internal_unique,add_history_attr
+from acrg.satellite.common import name_pressure_file, name_pressure_match,apply_filter,coord_order,define_pressure_levels,distance_lat,distance_lon,output_name,output,add_coords,ds_check_internal_unique,add_history_attr
 
 from acrg.satellite.gosat import latlon_filter, binned_mean, extract_files_dir_split
 
@@ -725,5 +725,91 @@ def oco2_process(site,species="co2",input_directory=input_directory,start=None,e
     return oco2       
     
 
-                 
-        
+def extract_files_oco2(directory, search_str=None, start=None, end=None, date_separator='', day=True):
+    '''
+    Extract filenames from a directory based on a search string and/or a date range.
+
+    Args:
+        directory (str): Directory to search.
+        search_str (str/None, optional): Search string with wildcard(s) to filter files.
+        start (str/None, optional): Start date in "YY-MM-DD" format.
+        end (str/None, optional): End date in "YY-MM-DD" format (exclusive).
+        date_separator (str): Separator between year, month, and day in filenames.
+        day (bool): Whether to include day in the date format.
+
+    Returns:
+        list: List of filenames matching the criteria.
+    '''
+    # Validate directory
+    if not os.path.isdir(directory):
+        print(f"ERROR: Directory '{directory}' does not exist.")
+        return []
+
+    # Validate date formats
+    if start and start.find('-') == -1:
+        print('WARNING: Start date is not in the correct format (YY-MM-DD).')
+        return []
+    if end and end.find('-') == -1:
+        print('WARNING: End date is not in the correct format (YY-MM-DD).')
+        return []
+    if (start and not end) or (end and not start):
+        print('WARNING: Both start and end dates must be specified.')
+        return []
+
+    # Build search string
+    search_str_short = search_str
+    if search_str:
+        search_str = os.path.join(directory, search_str)
+    else:
+        search_str = os.path.join(directory, "*")
+
+    # Get all filenames matching the search string
+    filenames = glob.glob(search_str)
+    filenames.sort()
+
+    # Filter files by date range if start and end are provided
+    if start and end:
+        # Convert start and end to datetime64[D]
+        start_date = np.datetime64(f"20{start[:2]}-{start[3:5]}-{start[6:]}", "D")
+        end_date = np.datetime64(f"20{end[:2]}-{end[3:5]}-{end[6:]}", "D")
+
+        # Generate the date range
+        date_range = np.arange(start_date, end_date, dtype="datetime64[D]")
+        date_range = [str(date) for date in date_range]
+
+        # Adjust date range based on 'day' and 'date_separator'
+        if not day:
+            date_range = ['-'.join(date.split('-')[:2]) for date in date_range]
+        date_range = [date.replace('-', date_separator) for date in date_range]
+
+        print(f'Finding files in range: {start} - {end} in directory {directory} using search string {search_str_short}')
+
+        # Filter filenames based on the date range
+        files = []
+        for filename in filenames:
+            try:
+                # Build regex pattern for date extraction
+                if date_separator and day:
+                    d_sep = f"[{date_separator}]"
+                    re_str = r"\d{4}" + d_sep + r"\d{2}" + d_sep + r"\d{2}"  # e.g., 2012-01-01
+                elif date_separator and not day:
+                    d_sep = f"[{date_separator}]"
+                    re_str = r"\d{4}" + d_sep + r"\d{2}"  # e.g., 2012-01
+                elif day:
+                    re_str = r"\d{8}"  # e.g., 20120101
+                else:
+                    re_str = r"\d{6}"  # e.g., 201201
+
+                # Extract date from filename
+                match = re.search(re_str, filename)
+                if match:
+                    extracted_date = match.group()
+                    if extracted_date in date_range:
+                        files.append(filename)
+            except AttributeError:
+                pass
+    else:
+        print(f'Finding files in directory {directory} with search string {search_str_short}')
+        files = filenames
+
+    return files
